+++ library/src/jvmTest/kotlin/io/github/littlesurvival/fetch/FetchTest.kt
@@ -5,12 +5,12 @@ import io.github.littlesurvival.YamiboClient
 import io.github.littlesurvival.YamiboForum
 import io.github.littlesurvival.core.YamiboResult
 import io.github.littlesurvival.dto.page.FavoriteType
-import io.github.littlesurvival.dto.page.PollOption
 import io.github.littlesurvival.dto.value.FormHash
 import io.github.littlesurvival.dto.value.ForumId
 import io.github.littlesurvival.dto.value.PollOptionId
 import io.github.littlesurvival.dto.value.PostId
 import io.github.littlesurvival.dto.value.SearchId
+import io.github.littlesurvival.dto.value.TagId
 import io.github.littlesurvival.dto.value.ThreadId
 import kotlinx.coroutines.runBlocking
 import kotlin.test.Test
@@ -54,8 +54,8 @@ class FetchTest {
 
     @Test
     fun testThread(): Unit = runBlocking {
-        // 568055, 535612, 564532, 565033, 557223, 535057, 568285, 567394, 568356, 568493, 566241, 551879
-        val threadResult = client.fetchThreadById(ThreadId(551879), page = 1)
+        // 568055, 535612, 564532, 565033, 557223, 535057, 568285, 567394, 568356, 568493, 566241, 551879, 568644
+        val threadResult = client.fetchThreadById(ThreadId(568644), page = 1)
         if (threadResult is YamiboResult.Success) {
             println("post size : ${threadResult.value.posts.size}")
         }
@@ -103,5 +103,20 @@ class FetchTest {
             formHash
         )
         debugLog("votePoll", pollResult)
+    }
+
+    @Test
+    fun testExtractTag() = runBlocking {
+        //568644, 568611
+        val tagResult = client.fetchExtractTagsInThreadById(ThreadId(568611))
+
+        debugLog("Tags", tagResult)
+    }
+
+    @Test
+    fun testTagPage() = runBlocking {
+        val result = client.fetchTagPageById(TagId(20666))
+
+        debugLog("TagPage", result)
     }
 }